refactor: move storage into core
