+++ packages/core/src/data-structs/data-structs.ts
@@ -1,5 +1,6 @@
 import { randomUUID } from "@llamaindex/env";
 import type { UUID } from "../global";
+import { BaseNode } from "../schema";
 import { IndexStructType } from "./struct-type";
 
 export abstract class IndexStruct {
@@ -61,6 +62,51 @@ export class KeywordTable extends IndexStruct {
         },
         {} as Record<string, string[]>,
       ),
+      type: this.type,
+    };
+  }
+}
+
+export class IndexDict extends IndexStruct {
+  nodesDict: Record<string, BaseNode> = {};
+  type: IndexStructType = IndexStructType.SIMPLE_DICT;
+
+  addNode(node: BaseNode, textId?: string) {
+    const vectorId = textId ?? node.id_;
+    this.nodesDict[vectorId] = node;
+  }
+
+  toJson(): Record<string, unknown> {
+    const nodesDict: Record<string, unknown> = {};
+
+    for (const [key, node] of Object.entries(this.nodesDict)) {
+      nodesDict[key] = node.toJSON();
+    }
+
+    return {
+      ...super.toJson(),
+      nodesDict,
+      type: this.type,
+    };
+  }
+
+  delete(nodeId: string) {
+    delete this.nodesDict[nodeId];
+  }
+}
+
+export class IndexList extends IndexStruct {
+  nodes: string[] = [];
+  type: IndexStructType = IndexStructType.LIST;
+
+  addNode(node: BaseNode) {
+    this.nodes.push(node.id_);
+  }
+
+  toJson(): Record<string, unknown> {
+    return {
+      ...super.toJson(),
+      nodes: this.nodes,
       type: this.type,
     };
   }
