+++ packages/core/src/data-structs/index.ts
@@ -1,2 +1,8 @@
-export { IndexStruct, KeywordTable } from "./data-structs";
+export {
+  IndexDict,
+  IndexList,
+  IndexStruct,
+  KeywordTable,
+} from "./data-structs";
+export { jsonToIndexStruct } from "./json-to-index-struct";
 export { IndexStructType } from "./struct-type";
