+++ packages/core/src/schema/index.ts
@@ -1,5 +1,10 @@
 export * from "./node";
-export { FileReader, TransformComponent, type BaseReader } from "./type";
+export {
+  FileReader,
+  TransformComponent,
+  type BaseReader,
+  type StoredValue,
+} from "./type";
 export type { BaseOutputParser } from "./type/base-output-parser";
 export { EngineResponse } from "./type/engine–response";
 export * from "./zod";
