+++ packages/core/src/storage/kv-store/index.ts
@@ -1,7 +1,35 @@
-import { DEFAULT_COLLECTION } from "@llamaindex/core/global";
 import { fs, path } from "@llamaindex/env";
-import { exists } from "../FileSystem.js";
-import { BaseKVStore, type StoredValue } from "./types.js";
+import { IndexStruct, jsonToIndexStruct } from "../../data-structs";
+import { DEFAULT_COLLECTION, DEFAULT_NAMESPACE } from "../../global";
+import type { StoredValue } from "../../schema";
+import { BaseIndexStore } from "../index-store";
+
+async function exists(path: string): Promise<boolean> {
+  try {
+    await fs.access(path);
+    return true;
+  } catch {
+    return false;
+  }
+}
+
+export abstract class BaseKVStore {
+  abstract put(
+    key: string,
+    val: StoredValue,
+    collection?: string,
+  ): Promise<void>;
+  abstract get(key: string, collection?: string): Promise<StoredValue>;
+  abstract getAll(collection?: string): Promise<Record<string, StoredValue>>;
+  abstract delete(key: string, collection?: string): Promise<boolean>;
+}
+
+export abstract class BaseInMemoryKVStore extends BaseKVStore {
+  abstract persist(persistPath: string): void;
+  static fromPersistPath(persistPath: string): BaseInMemoryKVStore {
+    throw new Error("Method not implemented.");
+  }
+}
 
 export type DataType = Record<string, Record<string, StoredValue>>;
 
@@ -96,5 +124,47 @@ export class SimpleKVStore extends BaseKVStore {
 
   static fromDict(saveDict: DataType): SimpleKVStore {
     return new SimpleKVStore(saveDict);
+  }
+}
+
+export class KVIndexStore extends BaseIndexStore {
+  private _kvStore: BaseKVStore;
+  private _collection: string;
+
+  constructor(kvStore: BaseKVStore, namespace: string = DEFAULT_NAMESPACE) {
+    super();
+    this._kvStore = kvStore;
+    this._collection = `${namespace}/data`;
+  }
+
+  async addIndexStruct(indexStruct: IndexStruct): Promise<void> {
+    const key = indexStruct.indexId;
+    const data = indexStruct.toJson();
+    await this._kvStore.put(key, data, this._collection);
+  }
+
+  async deleteIndexStruct(key: string): Promise<void> {
+    await this._kvStore.delete(key, this._collection);
+  }
+
+  async getIndexStruct(structId?: string): Promise<IndexStruct | undefined> {
+    if (!structId) {
+      const structs = await this.getIndexStructs();
+      if (structs.length !== 1) {
+        throw new Error("More than one index struct found");
+      }
+      return structs[0];
+    } else {
+      const json = await this._kvStore.get(structId, this._collection);
+      if (json == null) {
+        return;
+      }
+      return jsonToIndexStruct(json);
+    }
+  }
+
+  async getIndexStructs(): Promise<IndexStruct[]> {
+    const jsons = await this._kvStore.getAll(this._collection);
+    return Object.values(jsons).map((json) => jsonToIndexStruct(json));
   }
 }
