+++ packages/llamaindex/src/indices/index.ts
@@ -1,6 +1,5 @@
+export * from "@llamaindex/core/indices";
 export * from "./BaseIndex.js";
-export * from "./IndexStruct.js";
-export * from "./json-to-index-struct.js";
 export * from "./keyword/index.js";
 export * from "./summary/index.js";
 export * from "./vectorStore/index.js";
