+++ packages/llamaindex/src/ingestion/IngestionCache.ts
@@ -1,13 +1,15 @@
 import type { BaseNode, TransformComponent } from "@llamaindex/core/schema";
 import { MetadataMode } from "@llamaindex/core/schema";
-import { createSHA256 } from "@llamaindex/env";
 import {
   docToJson,
   jsonSerializer,
   jsonToDoc,
-} from "../storage/docStore/utils.js";
-import { SimpleKVStore } from "../storage/kvStore/SimpleKVStore.js";
-import type { BaseKVStore } from "../storage/kvStore/types.js";
+} from "@llamaindex/core/storage/doc-store";
+import {
+  SimpleKVStore,
+  type BaseKVStore,
+} from "@llamaindex/core/storage/kv-store";
+import { createSHA256 } from "@llamaindex/env";
 
 const transformToJSON = (obj: TransformComponent) => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
