+++ packages/llamaindex/src/ingestion/strategies/DuplicatesStrategy.ts
@@ -1,5 +1,5 @@
 import { BaseNode, TransformComponent } from "@llamaindex/core/schema";
-import type { BaseDocumentStore } from "../../storage/docStore/types.js";
+import type { BaseDocumentStore } from "@llamaindex/core/storage/doc-store";
 
 /**
  * Handle doc store duplicates by checking all hashes.
